backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Waitress==2.1.2
pymongo==4.6.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.1.2
